microInt applied in fiats

See merge request deepanshutr/persistenceClient!324

+++ app/models/common/Serializable.scala
@@ -62,7 +62,7 @@ object Serializable {
 
   implicit val tradeActivityTemplateWrites: OWrites[TradeActivityTemplate] = Json.writes[TradeActivityTemplate]
 
-  case class BillOfLading(id: String, consigneeTo: String, vesselName: String, portOfLoading: String, portOfDischarge: String, shipperName: String, shipperAddress: String, notifyPartyName: String, notifyPartyAddress: String, dateOfShipping: Date, deliveryTerm: String, assetDescription: String, assetQuantity: Int, quantityUnit: String, declaredAssetValue: Int) extends AssetDocumentContent
+  case class BillOfLading(id: String, consigneeTo: String, vesselName: String, portOfLoading: String, portOfDischarge: String, shipperName: String, shipperAddress: String, notifyPartyName: String, notifyPartyAddress: String, dateOfShipping: Date, deliveryTerm: String, assetDescription: String, assetQuantity: Int, quantityUnit: String, declaredAssetValue: Double) extends AssetDocumentContent
 
   implicit val assetDocumentContentWrites: Writes[AssetDocumentContent] = {
     case billOfLading: BillOfLading => Json.toJson(billOfLading)(Json.writes[BillOfLading])
